Moving method to base class instead
