+++ cohort-engine/src/main/java/com/ibm/cohort/engine/DirectoryLibrarySourceProvider.java
@@ -59,6 +59,6 @@ public class DirectoryLibrarySourceProvider extends MultiFormatLibrarySourceProv
 				return FileVisitResult.CONTINUE;
 			}
 		});
-		FhirHelperBackupProvider.addClasspathFhirHelpers(sources);
+		addClasspathFhirHelpers(sources);
 	}
 }
