+++ cohort-engine/src/main/java/com/ibm/cohort/engine/FhirLibraryLibrarySourceProvider.java
@@ -78,7 +78,7 @@ public class FhirLibraryLibrarySourceProvider extends MultiFormatLibrarySourcePr
 				}
 			}
 		}
-		FhirHelperBackupProvider.addClasspathFhirHelpers(sources);
+		addClasspathFhirHelpers(sources);
 
 		return numLoaded;
 	}
