+++ cohort-engine/src/main/java/com/ibm/cohort/engine/ZipStreamLibrarySourceProvider.java
@@ -64,6 +64,6 @@ public class ZipStreamLibrarySourceProvider extends MultiFormatLibrarySourceProv
 			}
 		}
 
-		FhirHelperBackupProvider.addClasspathFhirHelpers(sources);
+		addClasspathFhirHelpers(sources);
 	}
 }
