+++ cohort-engine/src/test/java/com/ibm/cohort/engine/TestClasspathLibrarySourceProvider.java
@@ -24,6 +24,6 @@ public class TestClasspathLibrarySourceProvider extends MultiFormatLibrarySource
 			}
 			formats.put(LibraryFormat.forString(resource), is);
 		}
-		FhirHelperBackupProvider.addClasspathFhirHelpers(sources);
+		addClasspathFhirHelpers(sources);
 	}
 }
